+++ CHANGELOG.md
@@ -207,6 +207,7 @@
 - Moved file handling to async @robgietema
 - Check migrations and profile upgrades on startup @robgietema
 - Make it possible to overwrite models @robgietema
+- Register event handler for content rules @robgietema
 
 ### Documentation